% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_repo_infos.R
\name{fetch_repo_infos}
\alias{fetch_repo_infos}
\title{Fetch relevant GitHub repo information from GitHub API}
\usage{
fetch_repo_infos(owner, repo)
}
\arguments{
\item{owner}{Owner of the GitHub repository (user or organisation)}

\item{repo}{Name of the GitHub repository}
}
\value{
A named vector:
\itemize{
\item \code{language} (character)
\item \code{license} (character)
\item \code{version} (character)
\item \code{has_codemeta} (logical)
\item \code{has_citation} (logical)
\item \code{has_zenodo} (logical)
}
}
\description{
Fetch:
\itemize{
\item main language (based on linguist. Might be inaccurate if people didn't use a
\code{.gitattributes} file)
\item license
\item latest tagged version
\item presence of a \code{codemeta.json} file on the default branch
\item presence of a \code{CITATION} file on the default branch
\item presence of a \code{.zenodo.json} file on the default branch
}
}
\examples{
fetch_repo_infos("ropensci", "lightr")

}
